requirements.txt
Django
bs4
django-environ
mysqlclient
requests